Questão 01

+++ Lista4/Q1.py
@@ -3,6 +3,6 @@ Faça um Programa que leia um vetor de 5 números inteiros e mostre-os.
 '''
 numeros = []
 for i in range(1,6):
-    num = float(input(f'Digite o {i}ª número: '))
+    num = int(input(f'Digite o {i}ª número: '))
     numeros.append(num)
 print(f'Os números do vetor são: {numeros}')
